Support gradients with mismatched color and opacity stops
The Lottie specification allows for gradient fills where
the number of color stops and opacity stops do not match.

Adds the logic to parse and render gradient.

The test asset also added for testApp manual test and
baseline test

Task-number: QTBUG-139234

+++ src/lottie/qlottiegfill_p.h
@@ -55,7 +55,7 @@ protected:
     QLottieSpatialProperty m_endPoint;
     QLottieProperty<qreal> m_highlightLength;
     QLottieProperty<qreal> m_highlightAngle;
-    QHash<int, QLottieProperty4D<QVector4D>> m_colors;
+    QHash<qreal, QLottieProperty4D<QVector4D>> m_colors;
     QGradient *m_gradient = nullptr;
 };
